Renamed installApps to deploy, updated plugin logic to use deploy task and extension, refactored UndeployTask
