+++ src/main/groovy/io/openliberty/tools/gradle/LibertyMultiServerTasks.groovy
@@ -32,7 +32,6 @@ import io.openliberty.tools.gradle.tasks.InstallFeatureTask
 import io.openliberty.tools.gradle.tasks.InstallLibertyTask
 import io.openliberty.tools.gradle.tasks.UninstallFeatureTask
 import io.openliberty.tools.gradle.tasks.CleanTask
-import io.openliberty.tools.gradle.tasks.InstallAppsTask
 import io.openliberty.tools.gradle.tasks.AbstractServerTask
 import io.openliberty.tools.gradle.tasks.CompileJSPTask
 import io.openliberty.tools.gradle.tasks.arquillian.ConfigureArquillianTask
@@ -130,13 +129,6 @@ class LibertyMultiServerTasks extends LibertyTasks {
             }
         })
 
-        overwriteTask('deploy', DefaultTask, {
-            description 'Deploys a supported file to the Liberty server.'
-            logging.level = LogLevel.INFO
-            group 'Liberty'
-            dependsOn getTaskList('deploy')
-        })
-
         overwriteTask('undeploy', DefaultTask, {
             description 'Removes an application from the Liberty server.'
             logging.level = LogLevel.INFO
@@ -165,14 +157,14 @@ class LibertyMultiServerTasks extends LibertyTasks {
             dependsOn getTaskList('cleanDirs')
         })
 
-        overwriteTask('installApps', DefaultTask, {
+        overwriteTask('deploy', DefaultTask, {
             description "Copy applications generated by the Gradle project to a Liberty server's dropins or apps directory."
             logging.level = LogLevel.INFO
             group 'Liberty'
-            List<String> installAppsTasks = getTaskList('installApps')
-            dependsOn project.tasks.withType(War), installAppsTasks
+            List<String> deployTasks = getTaskList('deploy')
+            dependsOn project.tasks.withType(War), deployTasks
             outputs.upToDateWhen {
-                tasksUpToDate(installAppsTasks)
+                tasksUpToDate(deployTasks)
             }
         })
 
@@ -191,16 +183,16 @@ class LibertyMultiServerTasks extends LibertyTasks {
         addTaskRule('Pattern: libertyStart-<Server Name>', 'libertyStart', StartTask, {
             dependsOn 'libertyCreate-' + server.name
 
-            if (dependsOnApps(server)) dependsOn 'installApps-' + server.name
+            if (dependsOnApps(server)) dependsOn 'deploy-' + server.name
         })
 
         addTaskRule('Pattern: libertyRun-<Server Name>', 'libertyRun', RunTask, {
             dependsOn 'libertyCreate-' + server.name
 
-            if (dependsOnApps(server)) dependsOn 'installApps-' + server.name
+            if (dependsOnApps(server)) dependsOn 'deploy-' + server.name
         })
 
-        addTaskRule('Pattern: installApps-<Server Name>', 'installApps', InstallAppsTask, {
+        addTaskRule('Pattern: deploy-<Server Name>', 'deploy', DeployTask, {
             dependsOn 'libertyCreate-' + server.name, project.tasks.withType(War)
         })
 
@@ -220,7 +212,7 @@ class LibertyMultiServerTasks extends LibertyTasks {
 
         addTaskRule('Pattern: libertyRun-<Server Name>', 'libertyRun', RunTask, {
             dependsOn 'libertyCreate-' + server.name
-            if (dependsOnApps(server)) dependsOn 'installApps-' + server.name
+            if (dependsOnApps(server)) dependsOn 'deploy-' + server.name
         })
 
         addTaskRule('Pattern: libertyStatus-<Server Name>', 'libertyStatus', StatusTask, {})
@@ -237,10 +229,6 @@ class LibertyMultiServerTasks extends LibertyTasks {
 
         addTaskRule('Pattern: libertyJavaDump-<Server Name>', 'libertyJavaDump', JavaDumpTask, {})
 
-        addTaskRule('Pattern: deploy-<Server Name>', 'deploy', DeployTask, {
-            dependsOn 'libertyStart-' + server.name
-        })
-
         addTaskRule('Pattern: undeploy-<Server Name>', 'undeploy', UndeployTask, {
             dependsOn 'libertyStart-' + server.name
         })
@@ -248,7 +236,7 @@ class LibertyMultiServerTasks extends LibertyTasks {
         addTaskRule('Pattern: cleanDirs-<Server Name>', 'cleanDirs', CleanTask, {})
 
         addTaskRule('Pattern: configureArquillian-<Server Name>', 'configureArquillian', ConfigureArquillianTask, {
-            dependsOn 'installApps-' + server.name, 'processTestResources'
+            dependsOn 'deploy-' + server.name, 'processTestResources'
             skipIfArquillianXmlExists = project.arquillianConfiguration.skipIfArquillianXmlExists
             arquillianProperties = project.arquillianConfiguration.arquillianProperties
         })
